feat: track TipsetStateCache size with prometheus metrics
