+++ src/shim/executor.rs
@@ -5,6 +5,7 @@ use super::trace::ExecutionEvent;
 use crate::shim::{
     econ::TokenAmount, fvm_shared_latest::ActorID, fvm_shared_latest::error::ExitCode,
 };
+use crate::utils::get_size::{GetSize, vec_heap_size_with_fn_helper};
 use cid::Cid;
 use fil_actors_shared::fvm_ipld_amt::Amtv0;
 use fvm_ipld_blockstore::Blockstore;
@@ -131,6 +132,16 @@ pub enum Receipt {
     V2(Receipt_v2),
     V3(Receipt_v3),
     V4(Receipt_v4),
+}
+
+impl GetSize for Receipt {
+    fn get_heap_size(&self) -> usize {
+        match self {
+            Self::V2(r) => r.return_data.bytes().get_heap_size(),
+            Self::V3(r) => r.return_data.bytes().get_heap_size(),
+            Self::V4(r) => r.return_data.bytes().get_heap_size(),
+        }
+    }
 }
 
 impl PartialEq for Receipt {
@@ -331,6 +342,19 @@ impl ActorEvent {
 pub enum StampedEvent {
     V3(StampedEvent_v3),
     V4(StampedEvent_v4),
+}
+
+impl GetSize for StampedEvent {
+    fn get_heap_size(&self) -> usize {
+        match self {
+            Self::V3(e) => vec_heap_size_with_fn_helper(&e.event.entries, |e| {
+                e.key.get_heap_size() + e.value.get_heap_size()
+            }),
+            Self::V4(e) => vec_heap_size_with_fn_helper(&e.event.entries, |e| {
+                e.key.get_heap_size() + e.value.get_heap_size()
+            }),
+        }
+    }
 }
 
 impl From<StampedEvent_v3> for StampedEvent {
