+++ src/utils/misc/mod.rs
@@ -8,6 +8,8 @@ pub use adaptive_value_provider::*;
 mod logo;
 pub use logo::*;
 pub mod env;
+mod r#type;
+pub use r#type::*;
 
 #[derive(Debug, Clone, PartialEq, Eq, strum::EnumString)]
 #[strum(serialize_all = "kebab-case")]
